+++ src/main/java/roomescape/reservation/dto/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.reservation.dto;
 
 public class ReservationResponse {
     private Long id;
