+++ src/main/java/roomescape/time/controller/TimeController.java
@@ -1,4 +1,4 @@
-package roomescape.time;
+package roomescape.time.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -8,6 +8,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.time.domain.AvailableTime;
+import roomescape.time.domain.Time;
+import roomescape.time.service.TimeService;
 
 import java.net.URI;
 import java.util.List;
